remove extra qualification
git-svn-id: http://kazuho.31tools.com/svn/q4m/trunk@119 23330352-1942-0410-abcd-d9b8ebed9e9b

+++ src/queue_cond.h
@@ -265,7 +265,7 @@ public:
   };
   struct pow_func : public binary_op<pow_func> {
     value_t bop(const value_t &x, const value_t &y) const {
-      return value_t::int_value(powl(x.l, y.l));
+      return value_t::int_value(static_cast<long long>(powl(x.l, y.l)));
     }
   };
